supersplat-rip/requirements.txt
requests
numpy
pillow
